requirements.txt
onnxruntime-gpu
pandas
opencv-python
pillow